Port complete statement tests to the new framework
